// Qatar2022-frontend/src/app/Models/spectator.ts
export interface Spectator{
    id?: number;
    firstName: string;
    lastName: string;
    email : string;
    country: string;
    tel: string;
    dob: Date;
    username: string;
    password: string;
    image:string;
    cardNumber: string;
    expiration: Date;
    cvv: string;
    tickets:any;
}